start position in Long instead on double

+++ tvplayer/src/main/java/com/kaltura/tvplayer/MediaOptions.java
@@ -9,7 +9,7 @@ import java.util.List;
 
 public abstract class MediaOptions {
     public String ks;
-    public double startPosition;
+    public Long startPosition;
     public List<PKExternalSubtitle> externalSubtitles;
 
     public abstract MediaEntryProvider buildMediaProvider(String serverUrl, int partnerId, String ks, String referrer);
